+++ addon/mixins/queryable-adapter.js
