+++ tests/dummy/app/models/post.js
@@ -1,7 +1,7 @@
 import DS from 'ember-data';
-import HasManyQuery from 'ember-data-has-many-query';
+import QueryableModelMixin from 'ember-data-has-many-query/mixins/queryable-model';
 
-export default DS.Model.extend(HasManyQuery.ModelMixin, {
+export default DS.Model.extend(QueryableModelMixin, {
   text: DS.attr('string'),
   comments: DS.hasMany('comment', {async: true})
 });
